Tools: Fix format error

+++ tools/grib_tools.cc
@@ -343,7 +343,7 @@ static int grib_tool_without_orderby(grib_runtime_options* options)
 #endif
             err = fseeko(infile->file, options->infile_offset, SEEK_SET);
             if (err) {
-                fprintf(stderr, "%s: Invalid file offset: %lld\n", tool_name, options->infile_offset);
+                fprintf(stderr, "%s: Invalid file offset: %ld\n", tool_name, (long)options->infile_offset);
                 exit(1);
             }
         }
